wait until the vr runtime is ready before show first frame
for openxr, if a frame is retrieved while the runtime is not ready, that
frame will never see the light of day. this usually means the splash
screen is never shown when using openxr runtime.

and this also works around a monado bug. because the vr runtime wasn't
ready when we tried to show the splash screen, we didn't get a chance to
save a view pose. which means after this frame, but before the first
camera frame is available, we will call `refresh()`. since `refresh()`
does not have a saved pose, it will submit 0 layers. and because of a
bug in monado, submitting 0 layers is not throttled. which means we will
be submitting repeatedly in rapid succession. and every time we submit,
the predictted display time is pushed into the future by one frame. when
the camera frame eventually arrives, the display time is usually a
couple of minutes into the future. basically means we will hang for
several minutes showing nothing.

making sure the runtime is ready, means we will have a saved pose right
away, which avoids this situation.